be-flask-api/requirements.txt
Flask==2.3.2
google-cloud-storage==2.5.0
sqlalchemy==2.0.16
mysqlclient==2.1.1
Werkzeug==2.3.6
